requirements.txt
Django
django-ninja
djangorestframework
djangorestframework-simplejwt
django-ninja-jwt
django-cors-headers
uvicorn
psycopg2
deepface